+++ worldspace/Frame.h
@@ -67,13 +67,13 @@ struct Frame {
 
 	/**
 	 * draw()
-	 * @brief Draws this frame to the console at it's origin point.
+	 * @brief Draws this frame to the console at it's origin polong.
 	 */
 	void draw()
 	{
 		// use dual-iterators to iterate both the frame, and console position from the origin offset
-		for (int consoleY{ _origin._y }, frameY{ 0 }; consoleY < _origin._y + static_cast<int>(_frame.size()); consoleY++, frameY++) {
-			for (int consoleX = _origin._x, frameX{ 0 }; consoleX < _origin._x + static_cast<int>(_frame.at(frameY).size()); consoleX++, frameX++) {
+		for (long consoleY{ _origin._y }, frameY{ 0 }; consoleY < _origin._y + static_cast<long>(_frame.size()); consoleY++, frameY++) {
+			for (long consoleX = _origin._x, frameX{ 0 }; consoleX < _origin._x + static_cast<long>(_frame.at(frameY).size()); consoleX++, frameX++) {
 				printf("%s%c", term::setCursorPosition(consoleX * 2, consoleY).c_str(), _frame.at(frameY).at(frameX));
 				if (_space_columns)
 					printf(" ");
